+++ database/migrations/2019_08_26_151021_add_category_id_to_courses_table.php
@@ -14,7 +14,7 @@ class AddCategoryIdToCoursesTable extends Migration
     public function up()
     {
         Schema::table('courses', function (Blueprint $table) {
-            $table->integer('category_id')->unsigned()->nullable();
+            $table->bigInteger('category_id')->unsigned()->nullable();
             $table->foreign('category_id')->references('id')->on('categories');
         });
     }
